Renaming locura
