+++ packages/service-integration/src/service_integration/cli.py
@@ -3,7 +3,7 @@
 import click
 
 from . import __version__
-from .commands import metadata, run_creator, update_compose_labels
+from .commands import compose_spec, metadata, run_creator
 
 
 @click.group()
@@ -13,6 +13,6 @@ def main():
 
 
 main.add_command(run_creator.main, "run-creator")
-main.add_command(update_compose_labels.main, "update-compose-labels")
+main.add_command(compose_spec.main, "compose")
 main.add_command(metadata.bump_version, "bump-version")
 main.add_command(metadata.get_version, "get-version")
